requirements.txt
discord.py>=2.0.0
anthropic>=0.3.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
pillow>=8.0.0
moviepy>=1.0.0
apscheduler>=3.9.0 